fix boolean bug (was it a pandas update that changed it?)

+++ psix/solo2psi.py
@@ -161,9 +161,9 @@ def get_psi_table_solo(intron_mtx_exons, minJR=5, minCell=20, tenX = False):
     SE_table = intron_mtx_exons.loc[se_events]
     SE_table.index = events
     
-    I1_filt = I1_table.index[(I1_table > minJR).sum(axis=1) > minCell]
-    I2_filt = I2_table.index[(I2_table > minJR).sum(axis=1) > minCell]
-    SE_filt = SE_table.index[(SE_table > minJR).sum(axis=1) > minCell]
+    I1_filt = I1_table.index[(I1_table > minJR).astype(int).sum(axis=1) > minCell]
+    I2_filt = I2_table.index[(I2_table > minJR).astype(int).sum(axis=1) > minCell]
+    SE_filt = SE_table.index[(SE_table > minJR).astype(int).sum(axis=1) > minCell]
     
     filtered_events = I1_filt.intersection(I2_filt).intersection(SE_filt)
